Break path resolvers out into their own classes.
