+++ AdvancedDLSupport/AdvancedDLSupport.csproj
@@ -19,4 +19,7 @@
     <PackageReference Include="StyleCop.Analyzers" Version="1.1.0-beta004" />
     <PackageReference Include="System.Reflection.Emit" Version="4.3.0" />
   </ItemGroup>
+  <ItemGroup>
+    <Folder Include="PathResolver" />
+  </ItemGroup>
 </Project>
